Add files via upload
Added a high level view of flow, then more detailed based on ranked operations on both sides. Need to define the steps in the middle. Will be partly based on the outcomes of other teams. Also, we need tenets to help us decide what we WILL define and what we WON'T define.